Fix dashboard spinner and moved dashboard into folder and added dashboard stylesheet
