+++ client/src/components/Dashboard/dashboard.style.css
@@ -1,0 +1,3 @@
+.main {
+  margin-top: 5em;
+}
